ml-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.4
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4
joblib==1.3.2
textblob==0.18.0
nltk==3.8.1
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
